move the adapters to their original location within the api logic
