+++ src/ensembl/production/metadata/api/adaptors/base.py
