+++ src/ensembl/production/metadata/api/adaptors/release.py
@@ -18,7 +18,7 @@ import sqlalchemy as db
 
 from ensembl.production.metadata.api.models import EnsemblRelease, EnsemblSite, GenomeRelease, Genome, GenomeDataset, \
     Dataset, ReleaseStatus
-from ensembl.production.metadata.grpc.adaptors.base import check_parameter, BaseAdaptor, cfg
+from ensembl.production.metadata.api.adaptors.base import check_parameter, BaseAdaptor, cfg
 
 logger = logging.getLogger(__name__)
