+++ src/tests/conftest.py
@@ -25,8 +25,8 @@ from grpc_reflection.v1alpha import reflection
 from ensembl.production.metadata.api.factories.datasets import DatasetFactory
 from ensembl.production.metadata.api.factories.genomes import GenomeFactory
 from ensembl.production.metadata.grpc import ensembl_metadata_pb2
-from ensembl.production.metadata.grpc.adaptors.genome import GenomeAdaptor
-from ensembl.production.metadata.grpc.adaptors.release import ReleaseAdaptor
+from ensembl.production.metadata.api.adaptors import GenomeAdaptor
+from ensembl.production.metadata.api.adaptors import ReleaseAdaptor
 
 
 def pytest_configure(config: Config) -> None:
